+++ blockapps-rest/ReleaseNotes.md
@@ -6,6 +6,7 @@
 - Add babel-preset-minify, remove gulp-uglify
 
 #### Minor upgrades
+- Added Bloc /compile endpoint
 - Add getAccounts to get the account details
 
 ### Version 7.6.3